+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Pashmak Changelog
 
+## 0.2 (2020-11-24)
+
+#### Added
+- added module path system
+- added `$__file__` and `$__dir__` general variables
+
+#### Fixed
+- raise error when namespace or function name contains `.` character
+
 ## v0.1.1
 
 #### Fixed